fix(types): fix some test-app type errors
