+++ test-app/types/global.d.ts
@@ -1,1 +1,11 @@
 import '@glint/environment-ember-loose';
+
+declare global {
+  interface Assert {
+    deepArrayContains(
+      array: unknown[],
+      element: unknown,
+      message?: string,
+    ): void;
+  }
+}
